pyproject.toml
[project]
name = "agentle"
version = "v0.9.2"
description = "..."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocache>=0.12.3",
    "aiohttp>=3.11.18",
    "aiosqlite>=0.21.0",
    "async-lru>=2.0.5",
    "dill>=0.4.0",
    "duckdb>=1.3.2",
    "google-genai>=1.14.0",
    "html-to-markdown>=2.3.4",
    "markitdown>=0.1.3",
    "mcp[cli]>=1.6.0",
    "pydantic>=2.11.3",
    "pypdf2>=3.0.1",
    "pyyaml>=6.0.2",
    "rock-solid-base>=0.1.14",
    "ujson>=5.10.0",
]

[[project.authors]]
name = "Arthur Brenno"
email = "64020210+arthurbrenno@users.noreply.github.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.directories]
packages = ["agentle"]

[tool.hatch.build.targets.wheel]
packages = ["agentle"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[dependency-groups]
cache = ["aiocache>=0.12.3"]
dev = [
    "aioboto3>=7.0.0",
 "aiocache>=0.12.3",
 "aiofiles>=24.1.0",
 "aiohttp>=3.11.18",
 "aiomysql>=0.2.0",
 "aioodbc>=0.5.0",
 "aiopg>=1.4.0",
 "aiosqlite>=0.21.0",
 "asyncmy>=0.2.10",
 "asyncpg>=0.30.0",
 "blacksheep>=2.2.0",
 "boto3>=1.37.37",
 "cassandra-driver>=3.29.2",
 "cerebras-cloud-sdk>=1.29.0",
 "couchbase>=4.3.5",
 "datamodel-code-generator>=0.35.0",
 "duckdb>=1.3.2",
 "duckduckgo-search>=8.0.4",
 "furo>=2024.8.6",
 "gemini-srt-translator>=1.8.5",
 "google-cloud-firestore>=2.21.0",
 "google-genai>=1.11.0",
 "groq>=0.22.0",
 "influxdb-client>=1.48.0",
 "ipykernel>=6.29.5",
 "jsonpickle>=4.1.1",
 "langfuse>=2.60.3",
 "mariadb>=1.1.12",
 "markdownify>=1.1.0",
 "motor>=3.7.0",
 "mypy>=1.15.0",
 "mypy-boto3-bedrock-runtime>=1.38.4",
 "mysql-connector-python>=9.3.0",
 "ollama>=0.5.1",
 "openai>=1.77.0",
 "openai-agents>=0.0.11",
 "openpyxl>=3.1.5",
 "oracledb>=3.1.0",
 "pandas>=2.2.3",
 "pikepdf>=9.7.0",
 "pillow>=10.1.0,<11.0.0",
 "playwright>=1.52.0",
 "psycopg>=3.2.7",
 "pydantic>=2.11.3",
 "pyinstrument>=5.0.1",
 "pymongo>=4.12.1",
 "pymssql>=2.3.4",
 "pymupdf>=1.25.5",
 "pyodbc>=5.2.0",
 "pypdf>=5.4.0",
 "pyright>=1.1.406",
 "pytest>=8.3.5",
 "pytest-asyncio>=0.23.8",
 "python-docx>=1.1.2",
 "python-dotenv>=1.1.0",
 "python-magic>=0.4.27",
 "python-pptx>=1.0.2",
 "qdrant-client>=1.15.0",
 "r2r>=3.5.14",
 "rarfile>=4.2",
 "redis>=6.0.0",
 "reportlab>=4.4.1",
 "rethinkdb>=2.4.10.post1",
 "soundfile>=0.13.1",
 "sphinx>=8.2.3",
 "streamlit>=1.45.0",
 "striprtf>=0.0.29",
 "svg-path>=6.3",
 "uvicorn>=0.34.1",
 "weasyprint>=65.1",
]
dev-db-rel = ["aiomysql>=0.2.0", "aioodbc>=0.5.0", "aiopg>=1.4.0", "aiosqlite>=0.21.0", "asyncmy>=0.2.10", "asyncpg>=0.30.0", "cx-oracle>=8.3.0", "ibm-db>=3.2.6", "mariadb>=1.1.12", "mysql-connector-python>=9.3.0", "oracledb>=3.1.0", "psycopg>=3.2.7", "psycopg2-binary>=2.9.10", "pymssql>=2.3.4", "pymysql>=1.1.1", "pyodbc>=5.2.0"]
dev-rb-nosql = ["aiocouch>=3.0.1", "aioinflux>=0.9.0", "aiomcache>=0.8.2", "aioredis>=2.0.1", "cassandra-driver>=3.29.2", "couchbase>=4.3.5", "influxdb-client>=1.48.0", "motor>=3.7.0", "pylibmc>=1.6.3", "pymongo>=4.12.1", "python-memcached>=1.62", "redis>=6.0.0", "rethinkdb>=2.4.10.post1"]
google = ["google-genai>=1.11.0"]
http = ["blacksheep>=2.2.0"]
langfuse = ["langfuse>=2.60.3"]
streamlit = ["streamlit>=1.45.0"]
